+++ frontend/app/src/pages/People/Tabs/Wanted.tsx
@@ -3,8 +3,8 @@ import { observer } from 'mobx-react-lite';
 import { BountyModal } from 'people/main/BountyModal';
 import { widgetConfigs } from 'people/utils/constants';
 import NoneSpace from 'people/utils/noneSpace';
-import { PostBounty } from 'people/widgetViews/postBounty';
-import WantedView from 'people/widgetViews/wantedView';
+import { PostBounty } from 'people/widgetviews/postbounty';
+import WantedView from 'people/widgetviews/wantedView';
 import React from 'react';
 import { Route, Switch, useHistory, useRouteMatch } from 'react-router-dom';
 import { useStores } from 'store';
