+++ frontend/app/src/pages/People/Tabs/index.tsx
@@ -1,6 +1,6 @@
 import { useIsMobile, usePerson } from 'hooks';
 import { observer } from 'mobx-react-lite';
-import RenderWidgets from 'people/widgetViews/renderWidgets';
+import RenderWidgets from 'people/widgetviews/renderWidgets';
 import { widgetConfigs } from 'people/utils/constants';
 import React, { useCallback, useEffect } from 'react';
 import { Route, Switch, useHistory, useLocation, useRouteMatch } from 'react-router-dom';
