+++ frontend/app/src/pages/Tickets/Tickets.tsx
@@ -1,8 +1,8 @@
 import { EuiGlobalToastList, EuiLoadingSpinner } from '@elastic/eui';
 import FirstTimeScreen from 'people/main/firstTimeScreen';
 import PageLoadSpinner from 'people/utils/pageLoadSpinner';
-import BountyHeader from 'people/widgetViews/bountyHeader';
-import WidgetSwitchViewer from 'people/widgetViews/widgetSwitchViewer';
+import BountyHeader from 'people/widgetviews/bountyHeader';
+import WidgetSwitchViewer from 'people/widgetviews/widgetSwitchViewer';
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import styled from 'styled-components';
