+++ frontend/app/src/people/widgetViews/renderWidgets.tsx
@@ -12,7 +12,7 @@ import { widgetConfigs } from '../utils/constants';
 import NoneSpace from '../utils/noneSpace';
 import OfferView from './offerView';
 import { AboutView } from './aboutView';
-import { PostBounty } from './postBounty';
+import { PostBounty } from './postbounty';
 import UserTickets from './userTicketsView';
 
 interface PanelProps {
@@ -107,15 +107,15 @@ function RenderWidgets({ widget }: RenderWidgetsProps) {
 
       const panelStyles = isMobile
         ? {
-            minHeight: 132
-          }
+          minHeight: 132
+        }
         : {
-            maxWidth: 291,
-            minWidth: 291,
-            marginRight: 20,
-            marginBottom: 20,
-            minHeight: 472
-          };
+          maxWidth: 291,
+          minWidth: 291,
+          marginRight: 20,
+          marginBottom: 20,
+          minHeight: 472
+        };
 
       fullSelectedWidget &&
         fullSelectedWidget.forEach((s: any, i: number) => {
@@ -127,9 +127,9 @@ function RenderWidgets({ widget }: RenderWidgetsProps) {
           const conditionalStyles =
             !isMobile && s?.paid
               ? {
-                  border: '1px solid #dde1e5',
-                  boxShadow: 'none'
-                }
+                border: '1px solid #dde1e5',
+                boxShadow: 'none'
+              }
               : {};
 
           elementArray.push(
