+++ frontend/app/src/people/widgetViews/wantedViews/mobileView.tsx
@@ -131,10 +131,9 @@ function MobileView(props: any) {
                   onClick={(e: any) => {
                     e.stopPropagation();
                     window.open(
-                      `/p/${
-                        {
-                          ...assignee
-                        }.owner_pubkey
+                      `/p/${{
+                        ...assignee
+                      }.owner_pubkey
                       }?widget=wanted`,
                       '_blank'
                     );
