+++ frontend/app/src/people/widgetViews/widgetSwitchViewer.tsx
@@ -1,9 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { observer } from 'mobx-react-lite';
-import OfferView from '../widgetViews/offerView';
-import WantedView from '../widgetViews/wantedView';
-import PostView from '../widgetViews/postView';
 import { useIsMobile } from '../../hooks';
 import { useStores } from '../../store';
 import { widgetConfigs } from '../utils/constants';
@@ -12,6 +9,9 @@ import NoResults from '../utils/noResults';
 import { uiStore } from '../../store/ui';
 import { bountyHeaderFilter, bountyHeaderLanguageFilter } from '../utils/filterValidation';
 import { colors } from '../../config/colors';
+import OfferView from './offerView';
+import WantedView from './wantedView';
+import PostView from './postView';
 import DeleteTicketModal from './deleteModal';
 
 interface PanelProps {
@@ -28,7 +28,7 @@ const Panel = styled.div<PanelProps>`
   border-bottom: ${(p: any) => (p.isMobile ? `2px solid ${p.color.grayish.G700}` : 'none')};
   :hover {
     box-shadow: ${(p: any) =>
-      p.isAssignee ? `0px 1px 6px ${p.color.black100}` : 'none'} !important;
+    p.isAssignee ? `0px 1px 6px ${p.color.black100}` : 'none'} !important;
   }
   :active {
     box-shadow: none !important;
@@ -78,16 +78,16 @@ function WidgetSwitchViewer(props: any) {
 
   const panelStyles = isMobile
     ? {
-        minHeight: 132
-      }
+      minHeight: 132
+    }
     : {
-        minWidth: '1100px',
-        maxWidth: '1100px',
-        marginBottom: 16,
-        borderRadius: '10px',
-        display: 'flex',
-        justifyContent: 'center'
-      };
+      minWidth: '1100px',
+      maxWidth: '1100px',
+      marginBottom: 16,
+      borderRadius: '10px',
+      display: 'flex',
+      justifyContent: 'center'
+    };
 
   const { peoplePosts, peopleWanteds, peopleOffers } = main;
 
@@ -158,9 +158,9 @@ function WidgetSwitchViewer(props: any) {
         const { person, body } = item;
         const conditionalStyles = body?.paid
           ? {
-              border: isMobile ? `2px 0 0 0 solid ${color.grayish.G600}` : '',
-              boxShadow: 'none'
-            }
+            border: isMobile ? `2px 0 0 0 solid ${color.grayish.G600}` : '',
+            boxShadow: 'none'
+          }
           : {};
 
         // if this person has entries for this widget
